[memset] fixed type so std::min don't complain

+++ src/goto-symex/builtin_functions.cpp
@@ -1023,7 +1023,7 @@ static inline expr2tc gen_value_by_byte(
 
       type2tc current_member_type = result->datatype_members[i]->type;
 
-      size_t current_member_size =
+      uint64_t current_member_size =
         type_byte_size(current_member_type).to_uint64();
 
       // Skip offsets
